run.sh: restart services on request conditionally

+++ run.sh
@@ -1,8 +1,5 @@
 #!/bin/bash
 
-kubectl delete -f mongo.yaml
-kubectl delete -f rabbit.yaml
-
 NS_DIR=./namespaces/*.yaml
 
 if [ "$(uname)" == "Darwin" ]; then
@@ -31,29 +28,38 @@ CATAMEL_REPO=$DOCKER_REPO/catamel
 
 echo $CATANIE_REPO
 
-helm del --purge local-mongodb 2> /dev/null
-helm del --purge local-rabbit 2> /dev/null
-helm del --purge local-node 2> /dev/null
-if [[ "KAFKA" -eq "1" ]]; then
-  helm del --purge local-kafka 2> /dev/null
-fi
-sleep 5 # or sync? let it purge the data before creating new ones
+answer=
+[ "$1" = "nopause" ] || \
+  read -p "Skip restarting base services (mongodb, rabbit, node)? [yN] " answer
+if [ "$answer" != "y" ]; then
+  kubectl delete -f mongo.yaml
+  kubectl delete -f rabbit.yaml
 
-for file in $NS_DIR; do
-  f="$(basename $file)"
-  ns="${f%.*}"
-  kubectl create -f $file
-  export LOCAL_ENV="$ns"
-  kubectl apply -f mongo.yaml
-  helm install stable/mongodb --version 0.4.15 --namespace $LOCAL_ENV --name local-mongodb
+  helm del --purge local-mongodb 2> /dev/null
+  helm del --purge local-rabbit 2> /dev/null
+  helm del --purge local-node 2> /dev/null
   if [[ "KAFKA" -eq "1" ]]; then
-    helm repo add incubator http://storage.googleapis.com/kubernetes-charts-incubator
-    helm install --name local-kafka incubator/kafka --namespace $LOCAL_ENV
+    helm del --purge local-kafka 2> /dev/null
   fi
-  kubectl apply -f rabbit.yaml
-  helm install stable/rabbitmq --version 0.6.3 --namespace $LOCAL_ENV --name local-rabbit --set rabbitmqUsername=admin,rabbitmqPassword=admin
-  helm install services/node-red --namespace $LOCAL_ENV --name local-node
-done
+  sleep 5; sync # let it purge the data before creating new ones
+
+  for file in $NS_DIR; do
+    f="$(basename $file)"
+    ns="${f%.*}"
+    kubectl create -f $file
+    export LOCAL_ENV="$ns"
+    kubectl apply -f mongo.yaml
+    helm install stable/mongodb --version 0.4.15 --namespace $LOCAL_ENV --name local-mongodb
+    if [[ "KAFKA" -eq "1" ]]; then
+      helm repo add incubator http://storage.googleapis.com/kubernetes-charts-incubator
+      helm install --name local-kafka incubator/kafka --namespace $LOCAL_ENV
+    fi
+    kubectl apply -f rabbit.yaml
+    helm install stable/rabbitmq --version 0.6.3 --namespace $LOCAL_ENV --name local-rabbit --set rabbitmqUsername=admin,rabbitmqPassword=admin
+    helm install services/node-red --namespace $LOCAL_ENV --name local-node
+  done
+fi
+
 export KUBE_NAMESPACE=yourns
 ./secret.sh "$KUBE_NAMESPACE"
 
@@ -61,11 +67,11 @@ export KUBE_NAMESPACE=yourns
 
 SERVICES_DIR=./services/*/*.sh
 
-answer=""
 for file in $SERVICES_DIR; do
+  answer=
   [ "$1" = "nopause" ] || \
-      read -p "Press <enter> to deploy the next service: $file (any input skips)" answer
-  [ -z "$answer" ] || continue
+    read -p "Skip restarting $file? [yN] " answer
+  [ "$answer" = "y" ] && continue
   cmd="bash $file $LOCAL_IP $CATANIE_REPO $CATAMEL_REPO $DOCKER_REPO/fs $DOCKER_REPO/ls"
   echo "Calling now $file:"
   echo " => $cmd"
